feat(slo): Remember last sort, view and group option used (#205340)
(cherry picked from commit 750a0f305823aff23cf9cc8b922fd3c07b63c622)
