+++ x-pack/solutions/observability/plugins/slo/public/embeddable/slo/overview/group_view/group_view.tsx
@@ -9,16 +9,14 @@ import { Filter } from '@kbn/es-query';
 import React, { useEffect, useState } from 'react';
 import { Subject } from 'rxjs';
 import { GroupView } from '../../../../pages/slos/components/grouped_slos/group_view';
-import { GroupByField } from '../../../../pages/slos/components/slo_list_group_by';
-import { SLOView } from '../../../../pages/slos/components/toggle_slo_view';
-import { SortField } from '../../../../pages/slos/hooks/use_url_search_state';
+import type { ViewType, GroupByField, SortField } from '../../../../pages/slos/types';
 import { buildCombinedKqlQuery } from './helpers/build_kql_query';
 
 interface Props {
   groupBy: GroupByField;
   groups?: string[];
   kqlQuery?: string;
-  view: SLOView;
+  view: ViewType;
   sort?: SortField;
   filters?: Filter[];
   reloadSubject: Subject<boolean>;
