+++ x-pack/solutions/observability/plugins/slo/public/embeddable/slo/overview/group_view/helpers/build_kql_query.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { GroupByField } from '../../../../../pages/slos/components/slo_list_group_by';
+import type { GroupByField } from '../../../../../pages/slos/types';
 import { buildCombinedKqlQuery } from './build_kql_query';
 
 describe('buildCombinedKqlQuery', () => {
