+++ x-pack/solutions/observability/plugins/slo/public/hooks/use_fetch_slo_groups.ts
@@ -18,7 +18,7 @@ import {
   DEFAULT_SLO_GROUPS_PAGE_SIZE,
   SLO_SUMMARY_DESTINATION_INDEX_PATTERN,
 } from '../../common/constants';
-import { GroupByField } from '../pages/slos/components/slo_list_group_by';
+import type { GroupByField } from '../pages/slos/types';
 import { SearchState } from '../pages/slos/hooks/use_url_search_state';
 import { useKibana } from './use_kibana';
 import { sloKeys } from './query_key_factory';
