+++ x-pack/solutions/observability/plugins/slo/public/pages/slos/components/common/sort_by_select.tsx
@@ -9,9 +9,10 @@ import { EuiPanel, EuiSelectableOption, EuiText } from '@elastic/eui';
 import { EuiSelectableOptionCheckedType } from '@elastic/eui/src/components/selectable/selectable_option';
 import { i18n } from '@kbn/i18n';
 import React, { useState } from 'react';
-import type { SortField, SearchState } from '../../hooks/use_url_search_state';
+import type { SearchState } from '../../hooks/use_url_search_state';
 import type { Option } from '../slo_context_menu';
 import { ContextMenuItem, SLOContextMenu } from '../slo_context_menu';
+import type { SortField } from '../../types';
 
 export interface Props {
   onStateChange: (newState: Partial<SearchState>) => void;
