+++ x-pack/solutions/observability/plugins/slo/public/pages/slos/components/grouped_slos/group_view.tsx
@@ -8,10 +8,8 @@ import { EuiEmptyPrompt, EuiFlexItem, EuiLoadingSpinner, EuiTablePagination } fr
 import { Filter } from '@kbn/es-query';
 import React, { useEffect } from 'react';
 import { useFetchSloGroups } from '../../../../hooks/use_fetch_slo_groups';
-import type { SortDirection } from '../../hooks/use_url_search_state';
-import { SortField, useUrlSearchState } from '../../hooks/use_url_search_state';
-import { GroupByField } from '../slo_list_group_by';
-import { SLOView } from '../toggle_slo_view';
+import { useUrlSearchState } from '../../hooks/use_url_search_state';
+import type { ViewType, GroupByField, SortDirection, SortField } from '../../types';
 import { SloGroupListEmpty } from './group_list_empty';
 import { SloGroupListError } from './group_list_error';
 import { GroupListView } from './group_list_view';
@@ -19,7 +17,7 @@ import { GroupListView } from './group_list_view';
 interface Props {
   groupBy: GroupByField;
   kqlQuery?: string;
-  view: SLOView;
+  view: ViewType;
   sort?: SortField;
   direction?: SortDirection;
   filters?: Filter[];
