+++ x-pack/solutions/observability/plugins/slo/public/pages/slos/components/grouped_slos/hooks/use_group_name.ts
@@ -9,7 +9,7 @@ import { i18n } from '@kbn/i18n';
 import { ALL_VALUE, GroupSummary } from '@kbn/slo-schema';
 import { assertNever } from '@kbn/std';
 import { SLI_OPTIONS } from '../../../../slo_edit/constants';
-import { GroupByField } from '../../slo_list_group_by';
+import type { GroupByField } from '../../../types';
 
 export function useGroupName(groupBy: GroupByField, group: string, summary?: GroupSummary) {
   const groupName = group.toLowerCase();
