+++ x-pack/solutions/observability/plugins/slo/public/pages/slos/components/toggle_slo_view.tsx
@@ -13,12 +13,12 @@ import React from 'react';
 import type { SearchState } from '../hooks/use_url_search_state';
 import { SLOSortBy } from './common/sort_by_select';
 import { SloGroupBy } from './slo_list_group_by';
-export type SLOView = 'cardView' | 'listView' | 'compactView';
+import type { ViewType } from '../types';
 
 interface Props {
-  onChangeView: (view: SLOView) => void;
+  onChangeView: (view: ViewType) => void;
   onStateChange: (newState: Partial<SearchState>) => void;
-  view: SLOView;
+  view: ViewType;
   state: SearchState;
   sloList?: FindSLOResponse;
   loading: boolean;
@@ -96,7 +96,7 @@ export function ToggleSLOView({
           })}
           options={toggleButtonsIcons}
           idSelected={view}
-          onChange={(id) => onChangeView(id as SLOView)}
+          onChange={(id) => onChangeView(id as ViewType)}
           isIconOnly
           isDisabled={loading}
         />
